add tests for more debug levels
